Move infer_spark_type
